Fix variable type to match returned type. No need for type-casting.

+++ SVPullToRefresh/UIScrollView+SVPullToRefresh.m
@@ -696,7 +696,7 @@ static char UIScrollViewPullToRefreshView;
         alphaGradient = CGGradientCreateWithColors(colorSpace, (__bridge CFArrayRef)alphaGradientColors, alphaGradientLocations);
     }else{
         const CGFloat * components = CGColorGetComponents([self.arrowColor CGColor]);
-        int numComponents = (int)CGColorGetNumberOfComponents([self.arrowColor CGColor]);
+        size_t numComponents = CGColorGetNumberOfComponents([self.arrowColor CGColor]);
         CGFloat colors[8];
         switch(numComponents){
             case 2:{
